requirements.txt
pandas~=2.2.3
numpy~=2.2.6
scikit-learn~=1.6.1
matplotlib~=3.10.6
seaborn
fuzzywuzzy~=0.18.0
python-Levenshtein
langdetect
sentence-transformers~=4.1.0
transformers~=4.52.4
torch~=2.8.0
jinja2~=3.1.6
aif360


joblib~=1.5.1
shap~=0.48.0
lime~=0.2.0.1
h11~=0.16.0
pillow~=11.2.1
pip~=25.2
typing_extensions~=4.14.0
attrs~=25.3.0
protobuf~=6.31.1
filelock~=3.18.0
setuptools~=80.9.0
regex~=2024.11.6
scipy~=1.15.3
click~=8.2.1
tqdm~=4.67.1
requests~=2.32.3
pyparsing~=3.2.3
pytz~=2025.2
Pygments~=2.19.2
colorama~=0.4.6
PyYAML~=6.0.2
sniffio~=1.3.1
idna~=3.10
charset-normalizer~=3.4.2
numba~=0.61.2
threadpoolctl~=3.6.0
mpmath~=1.3.0
sympy~=1.14.0
packaging~=24.2
networkx~=3.5
urllib3~=2.4.0
fsspec~=2025.5.1
shellingham~=1.5.4
ujson~=5.11.0
MarkupSafe~=3.0.2
cloudpickle~=3.1.1
pyarrow~=20.0.0
python-dateutil~=2.9.0.post0
plotly~=5.24.1
tenacity~=9.1.2
scikit-image~=0.25.2
certifi~=2025.4.26
tzdata~=2025.2
uvicorn~=0.34.3
websockets~=15.0.1
six~=1.17.0
Flask~=3.1.1
anyio~=4.9.0
pydantic~=2.11.5
annotated-types~=0.7.0
starlette~=0.46.2
python-multipart~=0.0.20
pydantic_core~=2.33.2
tornado~=6.5.1
watchdog~=6.0.0
uuid6~=2025.0.1
appdirs~=1.4.4
pycparser~=2.22
tokenizers~=0.21.1
Werkzeug~=3.1.3
safetensors~=0.5.3
huggingface-hub~=0.32.3
Levenshtein~=0.27.1
GitPython~=3.1.44
fastapi~=0.115.12
evidently
typing_extensions~=4.14.0
distro~=1.9.0
cryptography~=45.0.6
cffi~=1.17.1
nltk~=3.9.1
rich~=14.1.0
markdown-it-py~=4.0.0
Faker~=37.6.0
httpcore~=1.0.9
httpx~=0.28.1
mdurl~=0.1.2
patsy~=1.0.1
typer~=0.17.3
msgspec~=0.19.0
deprecation~=2.1.0
httptools~=0.6.4
watchfiles~=1.1.0
dynaconf~=3.2.11
rich-click~=1.8.9
litestar~=2.17.0
polyfactory~=2.22.2
multipart~=1.3.0
multidict~=6.6.4
pydantic_core~=2.33.2
iterative-telemetry~=0.0.10
typing-inspect~=0.9.0
statsmodels~=0.14.5
mypy_extensions~=1.1.0